+++ packages/hint-ie-flexbox-compat/src/meta.ts
@@ -1,0 +1,26 @@
+import { Category } from '@hint/utils-types';
+import { HintScope } from 'hint/dist/src/lib/enums/hint-scope';
+import { HintMetadata } from 'hint/dist/src/lib/types';
+
+import { getMessage } from './i18n.import';
+
+const meta: HintMetadata = {
+    docs: {
+        category: Category.compatibility,
+        description: getMessage('description', 'en'),
+        name: getMessage('name', 'en')
+    },
+    /* istanbul ignore next */
+    getDescription(language: string) {
+        return getMessage('description', language);
+    },
+    /* istanbul ignore next */
+    getName(language: string) {
+        return getMessage('name', language);
+    },
+    id: 'ie-flexbox-compat',
+    schema: [],
+    scope: HintScope.any
+};
+
+export default meta;
